pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ipub"
version = "0.1.0"
description = "Turn your technical notes into publishable content. CLI tool for scanning, drafting, reviewing and exporting."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "wbzuo"}]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-frontmatter>=1.0",
]

[project.optional-dependencies]
llm = ["anthropic>=0.30.0"]

[project.scripts]
ipub = "ipub.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
